Cleanup group activity page styling
Subtle changes were made to existing styling; change in color,
spacing between group sectionsi, headers, and thier contents.
Group description, number of annotations, and edit/leave
sections were combined into a single section- there's not a
compelling reason to keep them seperate and it makes styling
between sections easier if they aren't. Chose to not use
sidebar macro for the group_sidebar anymore as there are too
many differences to styling now. Added a new css class for
styling of new creator label that was added by open groups.
Using span, p, and a html tags instead of dl for formating the group
info. Also moved the description, group info, and edit/leave links
into the same section. Added div's for clustering and to enable
formating between categories within the sidebar section. This new
html tagging enables re-use of the same css for each section as
opposed to having to specify first and last variations. Note because
of the addition of the divs a new css styling theme was added called
group-info. The dl styling was also removed as it's no longer used.